Fix the form submit problem
This fixes the problem when you have a MaskMoney widget as last field of the form, enter a value and press the Enter key. The form is sent but the value you entered doesn't get updated in the hidden field.
So, making it blur when pressing Enter does save the value and the form is submitted with the hidden field updated.